requirements.txt
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
googleapis-common-protos==1.65.0
grpclib==0.4.7
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.8
multidict==6.0.5
mutagen==1.47.0
numpy==2.0.2
opencv-python==4.10.0.84
protobuf==5.28.0
pycryptodomex==3.20.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
viam-sdk==0.29.1
websockets==13.0.1
yt-dlp==2024.8.6

